rename paddlenlp_kernel
