+++ ops/src/paddlenlp_kernel/triton/cut_cross_entropy/linear_cross_entropy.py
