+++ ops/src/paddlenlp_kernel/triton/inf_cl/ring.py
