+++ ops/tests/triton/mamba/test_ssd.py
@@ -16,13 +16,13 @@ import paddle
 import paddle.nn.functional as F
 import pytest
 from einops import rearrange
-from ppnlp_kernel.triton.ssd_chunk_state import (
+from paddlenlp_kernel.triton.ssd_chunk_state import (
     _chunk_cumsum_fwd,
     _chunk_state_fwd,
     chunk_state,
     chunk_state_varlen,
 )
-from ppnlp_kernel.triton.ssd_state_passing import _state_passing_fwd
+from paddlenlp_kernel.triton.ssd_state_passing import _state_passing_fwd
 
 #######################################################################################################################################
 # patch paddle.allclose
